Conflicts:
	recipes/busybox/busybox_1.15.3.bb
	recipes/initscripts/initscripts_1.0.bb

+++ recipes/busybox/busybox-1.18.3/openpli/mdev-mount.sh
@@ -5,30 +5,37 @@ NOTIFYDEVNAME=$MDEV
 
 case "$ACTION" in
 	add|"")
+		ACTION="add"
+		# check if already mounted
+		if grep -q "^/dev/${MDEV} " /proc/mounts ; then
+			# Already mounted
+			exit 0
+		fi
 		# remove old mountpoint symlinks we might have for this device
 		rm -f $MOUNTPOINT
+		DEVBASE=`expr substr $MDEV 1 3`
+		# check for full-disk partition
+		if [ "${DEVBASE}" == "${MDEV}" ] ; then
+			if [ -d /sys/block/${DEVBASE}/${DEVBASE}1 ] ; then
+				# Partition detected, bail out!
+				exit 0
+			fi
+			if [ ! -f /sys/block/${DEVBASE}/size ] ; then
+				# No size at all
+				exit 0
+			fi
+			if [ `cat /sys/block/${DEVBASE}/size` == 0 ] ; then
+				# empty device, bail out
+				exit 0
+			fi
+		fi
 		# first allow fstab to determine the mountpoint
-		grep -q "/dev/$MDEV" /etc/fstab && mount /dev/$MDEV
+		mount /dev/$MDEV
 		if [ $? -ne 0 ] ; then
 			# no fstab entry, use automatic mountpoint
-			DEVBASE=`expr substr $MDEV 1 3`
-			# check for full-disk partition
-			if [ "${DEVBASE}" == "${MDEV}" ] ; then
-				if [ -d /sys/block/${DEVBASE}/${DEVBASE}1 ] ; then
-					# Partition detected, bail out!
-					exit 0
-				fi
-				if [ ! -f /sys/block/${DEVBASE}/size ] ; then
-					exit 0
-				fi
-				if [ `cat /sys/block/${DEVBASE}/size` == 0 ] ; then
-					# empty device, bail out
-					exit 0
-				fi
-			fi
 			REMOVABLE=`cat /sys/block/$DEVBASE/removable`
 			MODEL=`cat /sys/block/$DEVBASE/device/model`
-			if [ $REMOVABLE -eq "0" ]; then
+			if [ "${REMOVABLE}" -eq "0" ]; then
 				# mount the first non-removable device on /media/hdd
 				DEVICETYPE="hdd"
 			else
@@ -80,10 +87,6 @@ case "$ACTION" in
 			finddevice "cf"
 			finddevice "mmc1"
 		fi
-		# we don't really depend on the hotplug_e2_helper, but when it exists, call it
-		if [ -x /usr/bin/hotplug_e2_helper ]; then
-			/usr/bin/hotplug_e2_helper $ACTION /block/$MDEV $PHYSDEVPATH
-		fi
 		;;
 	remove)
 		umount /dev/$MDEV
@@ -92,7 +95,7 @@ case "$ACTION" in
 			DEVICETYPE="$1"
 			touch /dev/mdev.$DEVICETYPE
 			DEVSTATE=`cat /dev/mdev.$DEVICETYPE`
-			if [ $DEVSTATE == $MDEV ]; then
+			if [ "${DEVSTATE}" == "${MDEV}" ]; then
 				rm /dev/mdev.$DEVICETYPE
 				NOTIFYDEVNAME=$DEVICETYPE
 			fi
@@ -101,14 +104,19 @@ case "$ACTION" in
 		finddevice "usb"
 		finddevice "cf"
 		finddevice "mmc1"
-		# in case we had an automatic mountpoint, remove it
+		# remove automatic mountpoint (if not empty...)
 		rmdir $MOUNTPOINT
 		# in case it was a symlink, remove that as well
 		rm -f $MOUNTPOINT
 		rm -f /autofs/$MDEV
-		# we don't really depend on the hotplug_e2_helper, but when it exists, call it
-		if [ -x /usr/bin/hotplug_e2_helper ]; then
-			/usr/bin/hotplug_e2_helper $ACTION /block/$MDEV $PHYSDEVPATH
-		fi
+		;;
+	*)
+		# Unexpected keyword
+		exit 1
 		;;
 esac
+
+# we don't really depend on the hotplug_e2_helper, but when it exists, call it
+if [ -x /usr/bin/hotplug_e2_helper ]; then
+	/usr/bin/hotplug_e2_helper $ACTION /block/$MDEV $PHYSDEVPATH
+fi
